+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 When upgrading to a new version, make sure to follow the directions under the "Upgrading" header of the corresponding version.
 If there is no "Upgrading" header for that version, no post-upgrade actions need to be performed.
 
+## Upcoming
+
+### Bug Fixes
+- Fix a potential JSON decode error on failed requests
+
+
 ## 4.2 (2023-07-08)
 
 ### New Features
@@ -15,7 +21,6 @@ license
     - This is a false positive. Refer to
     [#102](https://github.com/jdholtz/auto-southwest-check-in/discussions/102)
     for more information
-
 
 ### Upgrading
 - The 'flights' key in the configuration file was renamed to 'reservations'. It is